LLM-RAG/code/requirements.txt
# Requirements for Generic Tokenizer with Embeddings
transformers>=4.20.0
torch>=1.12.0
PyYAML>=6.0
numpy>=1.21.0

# Optional dependencies for specific tokenizers
sentencepiece>=0.1.97  # For LLaMA/T5 tokenizers
protobuf>=3.20.0       # For LLaMA tokenizers
# tokenizers>=0.13.0     # Fast tokenizers
